requirements.txt
pyautogui
pillow
opencv-python
numpy
pyyaml